requirements.txt
djangorestframework==3.12.4
Django==3.1.7
django-environ==0.4.5
psycopg2==2.8.6
jdatetime==3.6.2
django-macaddress==1.8.0